fix: tree shaking и ssr
